PDFBOX-4892, sonar fix
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1909827 13f79535-47bb-0310-9956-ffa450edef68

+++ io/src/main/java/org/apache/pdfbox/io/IOUtils.java
@@ -128,7 +128,7 @@ public final class IOUtils
             }
             remaining -= bytesRead;
         }
-        return buffer.length - remaining;
+        return (long) buffer.length - remaining;
     }
 
     /**
